requirements.txt

--extra-index-url https://download.pytorch.org/whl/cpu

flask
torch==1.13.1
torchvision==0.14.1
numpy<2.0.0
opencv-python-headless
scipy
requests
boto3
python-dotenv
faiss-cpu
pymysql
cryptography==42.0.5